Introduce cloud_load_balancer component for first time.
